requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
ecdsa==0.14.1
faunadb==3.0.0
Flask==1.1.2
future==0.18.2
idna==2.10
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyasn1==0.4.8
python-jose==3.2.0
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
